Add head picture for register
1. Fix some bugs
2. Add head picture for register

+++ src/main/java/com/rain/gameCommunity/dao/GameDAO.java
@@ -12,5 +12,5 @@ public interface GameDAO {
 
 	public GameEntity queryGameByName(@Param("gameName") String gameName);
 
-	public GameEntity queryGameById(@Param("id") int id);
+	public GameEntity queryGameById(@Param("id") long id);
 }
